Migrated loading code to seperate file
